Task calibration example output object by Bayes

This is an example output object from running the fit.model.byaes function by using the Bayesian estimator option. It is a result of calibrating an oral reading fluency data set from Nese and Kamata (2014-2018) with xxxx students on 150 passages.

Usage

passage.calib.bayes

Format

two lists: $pass.param with 150 rows and $hyper.param with 4 variables

$pass.param

a

parameter controlling binomial success probabilities

b

parameter controlling binomial success probabilities

alpha

parameter controlling reading times

beta

parameter controlling reading times

se_a

standard error of a

se_b

standard error of b

se_alpha

standard error of alpha

se_beta

standard error of beta

passage.id

passage ID

nwords.p

the total

$hyper.param

vartau

variance of latent reading ability tau

rho

correlation between two latent variables

se_vartau

standard error of vartar

se_rho

standard error of rho

Source

https://jnese.github.io/core-blog/

References

Nese, J. F. T. & Kamata, A. (2014-2018). Measuring Oral Reading Fluency: Computerized Oral Reading Evaluation (Project No. R305A140203) [Grant]. Institute of Education Sciences, U.S. Department of Education. https://ies.ed.gov/funding/grantsearch/details.asp?ID=1492