Passage-level Oral Reading Fluency assessment data set

This is an example data set. It is a passage-level Oral Reading Fluency assessment data set for 85 students who were assigned to read 2 to 12 passages among the same 12 passages. The data is a small subset of the data collected by Nese and Kamata (2014-2018).

Usage

passage2

Format

A data frame with 847 rows and 7 variables.

id.student

unique student identifier

occasion

identifier for longitudinal assessment occasions; here a triannual assessment administered in the fall, winter, and spring of a school year

grade

student grade level

id.passage

unique passage identifier

numwords.pass

total number of words in the passage

wrc

words read correct

sec

seconds to read the passage

Source

https://jnese.github.io/core-blog/

References

Nese, J. F. T. & Kamata, A. (2014-2018). Measuring Oral Reading Fluency: Computerized Oral Reading Evaluation (Project No. R305A140203) [Grant]. Institute of Education Sciences, U.S. Department of Education. https://ies.ed.gov/funding/grantsearch/details.asp?ID=1492