This is an interface function to call and run scoring or bootstrap.

Copyright (C) 2021-2023 The ORF Project Team

Usage

scoring(
  calib.data = NA,
  person.data = NA,
  person.id = "",
  task.id = "",
  occasion = "",
  group = "",
  max.counts = "",
  obs.counts = "",
  time = "",
  cases = NULL,
  est = "map",
  se = "analytic",
  failsafe = 0,
  bootstrap = 100,
  external = NULL,
  type = "general"
)

Arguments

calib.data

- fit.model class object

person.data

- individual reading data

person.id

The column name in the data that represents the unique individual identifier.

task.id

The column name in the data that represents the unique task identifier.

occasion

The column name in the data that represents the unique occasion.

group

The column name in the data that represents the unique group.

max.counts

The column name in the data that represents the number of words in a task.

obs.counts

The column name in the data that represents the words read correctly for each case.

time

The column name in the data that represents the time, in seconds, for each case.

cases

- individual id vectors, will directly use task data if no calib.data provided

est

- estimator keyword / c("mle", "map", "eap", "mcmc"), default is mcmc

se

- standard error keyword / c("analytic", "bootstrap"), default is analytic

failsafe

- retry time for bootstrap / default 0, can set to 5 ~ 50

external

- if not NULL, will use not student read passages for estimating

type

- output type, "general" and "orf", default "general" only output tau & theta. "orf" will output wcpm

bootstrp

- set K number of bootstrap / default 100

Value

scoring list or Bootstrap dataset

Details

This program is free software; you can redistribute it and/or modify it under the terms of the GNU General Public License as published by the Free Software Foundation; either version 3 of the License, or (at your option) any later version. This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for more details. A copy of the GNU General Public License is available at http://www.gnu.org/licenses/